requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Cloud & Database
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
sqlalchemy==2.0.23

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2

# GenAI & LangChain (Gemini-powered)
langchain==0.0.350
langchain-community==0.0.10
langchain-google-genai==0.0.6
google-generativeai==0.3.2
langsmith==0.0.69
langchain-experimental==0.0.47

# Data Science & ML
scikit-learn==1.3.2
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2

# Time Series Analysis
statsmodels==0.14.1

# Big Data & Spark
pyspark==3.5.0

# Deep Learning
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# Statistical Testing & Experimentation
scipy==1.11.4
pingouin==0.5.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# PDF Generation & Formatting
reportlab==4.0.7
markdown==3.5.1